Add generic Offcanvas component.
