+++ src/components/cipp/TitleButton.js
@@ -1,20 +1,20 @@
 import React from 'react'
 import { CButton } from '@coreui/react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faPlus } from '@fortawesome/free-solid-svg-icons'
 import PropTypes from 'prop-types'
 
-export default function TitleButton({ icon = faPlus, title, href }) {
+export default function TitleButton({ icon, title, href }) {
+  console.log(icon)
   return (
     <CButton size="sm" color="primary" href={href}>
-      <FontAwesomeIcon icon={faPlus} className="pe-1" />
+      {icon ? <FontAwesomeIcon icon={icon} className="pe-1" /> : null}
       {title}
     </CButton>
   )
 }
 
 TitleButton.propTypes = {
-  icon: PropTypes.node,
+  icon: PropTypes.object,
   title: PropTypes.string.isRequired,
   href: PropTypes.string.isRequired,
 }
